requirements.txt
aiofiles==0.7.0
anyio==3.3.1
asgiref==3.4.1
attrs==21.2.0
Authlib==0.15.4
black==21.9b0
bleach==4.1.0
build==0.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
docutils==0.17.1
fastapi==0.68.1
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==23.2.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pavlok==0.1.1
pep517==0.11.0
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.10.0
pyclickup==0.1.4
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
readme-renderer==29.0
regex==2021.9.24
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
tqdm==4.62.3
twine==3.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
webencodings==0.5.1
zipp==3.5.0
